+++ lib/Datasource/ExternalFile.php
@@ -58,6 +58,7 @@ class ExternalFile implements IDatasource
     {
         $template = array();
         $template[] = ['id' => 'link', 'name' => $this->l10n->t('External URL'), 'placeholder' => 'url'];
+        $template[] = ['id' => 'hasHeader', 'name' => $this->l10n->t('Header row'), 'placeholder' => 'true-' . $this->l10n->t('Yes').'/false-'.$this->l10n->t('No'), 'type' => 'tf'];
         $template[] = ['id' => 'offset', 'name' => $this->l10n->t('Ignore leading rows'), 'placeholder' => $this->l10n->t('Number of rows'), 'type' => 'number'];
         $template[] = ['id' => 'columns', 'name' => $this->l10n->t('Select columns'), 'placeholder' => $this->l10n->t('e.g. 1,2,4 or leave empty'), 'type' => 'columnPicker'];
         return $template;
@@ -98,11 +99,17 @@ class ExternalFile implements IDatasource
         if (isset($option['columns']) && strlen($option['columns']) > 0) {
             $selectedColumns = str_getcsv($option['columns'], ',');
         }
+
+        // get the delimiter by reading the first row
         $delimiter = $this->detectDelimiter($rows[0]);
 
+        // the first row will define the column headers, even if it is not a real header
         $header = str_getcsv($rows[0], $delimiter);
-        //ToDo: option to have datasource without header
-        //$rows = array_slice($rows, 1);
+
+        // if the data has a real header, remove the first row
+        if (!isset($option['hasHeader']) or $option['hasHeader'] !== 'false') {
+            $rows = array_slice($rows, 1);
+        }
 
         $data = array();
         if (count($selectedColumns) !== 0) {