+++ lib/Datasource/File.php
@@ -63,6 +63,7 @@ class File implements IDatasource
     {
         $template = array();
         $template[] = ['id' => 'link', 'name' => $this->l10n->t('File'), 'placeholder' => $this->l10n->t('File'), 'type' => 'filePicker'];
+        $template[] = ['id' => 'hasHeader', 'name' => $this->l10n->t('Header row'), 'placeholder' => 'true-' . $this->l10n->t('Yes').'/false-'.$this->l10n->t('No'), 'type' => 'tf'];
         $template[] = ['id' => 'offset', 'name' => $this->l10n->t('Ignore leading rows'), 'placeholder' => $this->l10n->t('Number of rows'), 'type' => 'number'];
         $template[] = ['id' => 'columns', 'name' => $this->l10n->t('Select columns'), 'placeholder' => $this->l10n->t('e.g. 1,2,4 or leave empty'), 'type' => 'columnPicker'];
         return $template;
@@ -77,10 +78,7 @@ class File implements IDatasource
      */
     public function readData($option): array
     {
-        $data = array();
         $error = 0;
-        $selectedColumns = array();
-
         $file = $this->rootFolder->getUserFolder($option['user_id'])->get($option['link']);
         $rows = str_getcsv($file->getContent(), "\n");
 
@@ -89,14 +87,23 @@ class File implements IDatasource
             $rows = array_slice($rows, $option['offset']);
         }
 
+        $selectedColumns = array();
         if (isset($option['columns']) && strlen($option['columns']) > 0) {
             $selectedColumns = str_getcsv($option['columns'], ',');
         }
+
+        // get the delimiter by reading the first row
         $delimiter = $this->detectDelimiter($rows[0]);
 
+        // the first row will define the column headers, even if it is not a real header
         $header = str_getcsv($rows[0], $delimiter);
-        $rows = array_slice($rows, 1);
 
+        // if the data has a real header, remove the first row
+        if (!isset($option['hasHeader']) or $option['hasHeader'] !== 'false') {
+            $rows = array_slice($rows, 1);
+        }
+
+        $data = array();
         if (count($selectedColumns) !== 0) {
             // if only a subset of columns or fixed column values are set, they are replaced here
             $header = $this->minimizeRow($selectedColumns, $header);